requirements.txt
Django==1.8.1
django-bootstrap3==5.4.0
django-filter==0.10.0
djangorestframework==3.1.2
Markdown==2.6.2
Pygments==2.0.2